app/requirements.txt
pip==24.0
pytest==7.4.0
httpx==0.24.1
ipython==8.14.0
motor==3.3.0
mysql-connector-python==8.2.0
flask==3.0.0
bcrypt==4.1.2
Flask-Session==0.6.0
pyOpenSSL==24.0.0
Werkzeug==3.0.1
flask-login==0.6.3
Flask-Bcrypt==1.0.1
